+++ test/test_mixins/test_string_commands.py
@@ -9,6 +9,7 @@ import redis.client
 from redis.exceptions import ResponseError
 
 from fakeredis._helpers import current_time
+from .. import testtools
 from ..testtools import raw_command
 
 
@@ -265,7 +266,7 @@ def test_set_ex_overflow(r: redis.Redis):
 
 def test_set_px_overflow(r: redis.Redis):
     with pytest.raises(ResponseError):
-        r.set("foo", "bar", px=2**63 - 2)  # Overflows after adding current time
+        r.set("foo", "bar", px=2 ** 63 - 2)  # Overflows after adding current time
 
 
 def test_set_px(r: redis.Redis):
@@ -278,6 +279,7 @@ def test_set_px_using_timedelta(r: redis.Redis):
     assert r.get("foo") == b"bar"
 
 
+@testtools.run_test_if_redispy_ver('lt', '5.9')  # This will run for redis-py 4.2.0 or above.
 def test_set_conflicting_expire_options(r: redis.Redis):
     with pytest.raises(ResponseError):
         r.set("foo", "bar", ex=1, px=1)